packages/ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
transformers==4.35.0
torchaudio==2.1.0
numpy==1.24.0
scipy==1.11.0
librosa==0.10.0
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.6.0
